requirements.txt
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
grpcio==1.60.1
grpcio-tools==1.60.1
h11==0.14.0
idna==3.6
outcome==1.3.0.post0
overrides==7.7.0
protobuf==4.25.2
PySocks==1.7.1
python-dotenv==1.0.1
requests==2.31.0
robotframework==7.0
robotframework-assertion-engine==3.0.3
robotframework-browser==18.1.0
robotframework-pythonlibcore==4.4.0
robotframework-seleniumlibrary==6.2.0
selenium==4.19.0
setuptools==69.2.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.10.0
urllib3==2.2.1
wrapt==1.16.0
wsproto==1.2.0
